Move utils that should be kept out of remote_storage
